amanzi-soweto/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==6.0.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
bcc==0.18.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==7.0.2
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
chrome-gnome-shell==0.0.0
click==8.2.1
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.17
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
ipykernel==7.0.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
lark==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.7.2
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
mistune==3.1.4
more-itertools==8.10.0
multidict==6.7.1
narwhals==2.17.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.6.0
pop-transition==1.1.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydbus==0.6.0
pydeck==0.9.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.2.5
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-json-logger==4.0.0
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.37.0
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
sessioninstaller==0.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
streamlit==1.55.0
systemd-python==234
tenacity==9.1.4
terminado==0.18.1
texttable==1.6.4
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
twilio==9.10.2
types-python-dateutil==2.9.0.20251008
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
uri-template==1.3.0
urllib3==1.26.5
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xdg==5
xkit==0.0.0
yarl==1.23.0
zipp==1.0.0
